rename `dynamics_augmentation` file
