+++ openscvx/trajoptproblem.py
@@ -18,7 +18,7 @@ from openscvx.config import (
     Config,
 )
 from openscvx.dynamics import Dynamics
-from openscvx.augmentation.state_augmentation import build_augmented_dynamics
+from openscvx.augmentation.dynamics_augmentation import build_augmented_dynamics
 from openscvx.augmentation.ctcs import sort_ctcs_constraints
 from openscvx.constraints.violation import get_g_funcs, CTCSViolation
 from openscvx.discretization import get_discretization_solver
